split validator rules file
