+++ packages/editor/src/validator/utils.ts
@@ -1,0 +1,4 @@
+export function isExpression (str: unknown) {
+  const regExp = new RegExp('.*{{.*}}.*');
+  return typeof str === 'string' && regExp.test(str)
+}
